src/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.2.1
bcrypt==4.1.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-cli==0.0.4
frozendict==2.3.10
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
jwcrypto==1.5.6
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.6.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.4
pyuploadcare==5.1.0
PyYAML==6.0.1
replicate==0.25.1
requests==2.31.0
rich==13.7.1
rsa==4.9
secure==0.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions
urllib3==2.0.7
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
